Don't need to convert to int from Integer for AppdefEntityID

+++ src/org/hyperic/hq/ui/action/resource/common/monitor/visibility/CompareMetricsFormPrepareAction.java
@@ -6,7 +6,7 @@
  * normal use of the program, and does *not* fall under the heading of
  * "derived work".
  * 
- * Copyright (C) [2004, 2005, 2006], Hyperic, Inc.
+ * Copyright (C) [2004-2008], Hyperic, Inc.
  * This file is part of HQ.
  * 
  * HQ is free software; you can redistribute it and/or modify
@@ -33,11 +33,16 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.TreeMap;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.apache.struts.action.ActionForm;
+import org.apache.struts.action.ActionForward;
+import org.apache.struts.action.ActionMapping;
+import org.apache.struts.tiles.ComponentContext;
 import org.hyperic.hq.appdef.shared.AppdefEntityID;
 import org.hyperic.hq.bizapp.shared.MeasurementBoss;
 import org.hyperic.hq.bizapp.shared.uibeans.ResourceMetricDisplaySummary;
@@ -52,13 +57,6 @@ import org.hyperic.hq.ui.util.RequestUtils;
 import org.hyperic.hq.ui.util.SessionUtils;
 import org.hyperic.util.config.InvalidOptionException;
 import org.hyperic.util.units.FormattedNumber;
-
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.apache.struts.action.ActionForm;
-import org.apache.struts.action.ActionForward;
-import org.apache.struts.action.ActionMapping;
-import org.apache.struts.tiles.ComponentContext;
 
 public class CompareMetricsFormPrepareAction extends WorkflowPrepareAction {
     protected static Log log =
@@ -93,7 +91,7 @@ public class CompareMetricsFormPrepareAction extends WorkflowPrepareAction {
         AppdefEntityID[] entIds = new AppdefEntityID[rids.length];        
         for (int i = 0; i < rids.length; i++) {
             entIds[i] = new AppdefEntityID(cform.getAppdefType().intValue(),
-                                           rids[i].intValue());
+                                           rids[i]);
         }
 
         try {
